Create ConWin.py
The class can be used to debug code. This is possible through generating labels with lights (one ConWin instance) and call the check function at the needed place in code. If the light with the corresponding labels switches to green, the code reached until that line. Therefore, if a program crashes, ConWin makes it possible to check in which line the error really occurs.